matthew suggests a double cast so sf mortgages work right. :)

+++ lib/libc/time/difftime.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: difftime.c,v 1.9 2005/08/08 08:05:38 espie Exp $ */
+/*	$OpenBSD: difftime.c,v 1.10 2015/02/09 08:36:53 tedu Exp $ */
 /* This file is placed in the public domain by Ted Unangst. */
 
 #include "private.h"
@@ -6,5 +6,5 @@
 double
 difftime(time_t time1, time_t time0)
 {
-	return time1 - time0;
+	return time1 - (double)time0;
 }
